fix: resolve 8 compilation errors across app and server modules
- Add androidx.security:security-crypto dependency for DeviceTrustStore
- Remove non-existent timeout param from MessageReader.readMessage()
- Fix Short→Int type mismatch in FrameDecoder (FrameMessage.width/height)
- Add fileName/localPath to ActiveTransferState in FileTransferController
- Fix Byte or operator in ControllerService and Server (toInt conversion)
- Fix R.id mismatches: btn_refresh→btn_refresh_local,
  btn_select_file→btn_select_files, switch_enable_share→switch_share_enable,
  btn_save→btn_save_config

+++ app/src/main/kotlin/com/scrcpybt/app/codec/FrameDecoder.kt
@@ -55,10 +55,10 @@ class FrameDecoder {
             val pixels = codec.decode(msg) ?: return null
             if (pixels.isEmpty()) return null
 
-            val width = msg.width
-            val height = msg.height
+            val width = msg.width.toInt()
+            val height = msg.height.toInt()
 
-            // 尺寸匹配时复用 Bitmap
+            // 尺寸匹配时复用 Bitmap | Reuse Bitmap when dimensions match
             val bitmap = reusableBitmap
             if (bitmap == null || bitmap.width != width || bitmap.height != height) {
                 reusableBitmap = Bitmap.createBitmap(width, height, Bitmap.Config.ARGB_8888)
